Réorganisation des classes, ajout dun autoloader.php, refonte de toute les pages pour avoir uniquement lautoloader, Code ajouter, checkCode ajouter
